% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correction.R
\name{correct_rel_deviation}
\alias{correct_rel_deviation}
\title{Correcting Systematic Bias in Metabolomic Timecourse Data}
\usage{
correct_rel_deviation(time, concentration, metabolite,
  f_smooth = met_smooth_gam, max.iter = 20, min.deviation = 0.02, ...)
}
\arguments{
\item{time}{A vector of times or sample numbers for metabolite time-courses.
Note, there should be a time value for every \code{concentration}
for every \code{metabolite} e.g. time = c(1, 2, 3, 1, 2, 3),
concentration = c(20, 15, 10, 3, 6, 9),
metabolite = c('glc', 'glc', 'glc', 'lac', 'lac', 'lac').}

\item{concentration}{A vector of metabolite concentrations.}

\item{metabolite}{A vector of metabolite names that correspond to \code{time}
and \code{concentration}.}

\item{f_smooth}{Smoothing function (set to met_smooth_gam by default).}

\item{max.iter}{The algorithm correct the single most deviating point at one
time (as it may influence the identification of other
deviations). \code{max.iter} controls the maximum number
of correction passes.}

\item{min.deviation}{Smallest median relative deviation to identify as a
bias. 0.02 has been found to be a good default for
a diverse set of metabolic time-courses from
higher eukaryote cell culture.}

\item{...}{Arguments to be passed into \code{f_smooth}, (such as k or
span parameters for met_smooth_gam and met_smooth_loess
respectively). Note, the function attempts to identify the use of
met_smooth_gam and met_smooth_loess in order to provide reasonable
default smoothing parameters based on observations of metabolic
time-courses from higher eukaryote cell culture. It is highly
recommened to provide explicit smoothing parameters.}
}
\value{
A vector of corrected concentrations.
}
\description{
Identifies systematic deviations in metabolic data using a smoothing fit.
Timepoints that have a median relative deviation above a threshold value
are assumed to be influenced by a measurement or methodological bias as
compared to the overall trends metabolite concentrations.
}
\examples{
# TODO
}

