% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correction.R
\name{met_smooth_gam}
\alias{met_smooth_gam}
\title{GAM Smoothing}
\usage{
met_smooth_gam(time, concentration, new.time = NULL, ...)
}
\arguments{
\item{time}{Time or sample number for metabolite time-course.}

\item{concentration}{Metabolite concentration.}

\item{new.time}{Optional vector for new independent variables.}

\item{...}{Arguments to be passed into the s() smoothing function, such as
k -- the dimension of the basis used to represent the smooth term.}
}
\value{
A vector of smoothed concentrations.
}
\description{
Wraps GAM smoothing function for use in bias correction algorithm.
}
\examples{
# Simulating concave curve with 10 points
par <- c(3.5, 4.5, 2.5, 3.5, 0.0, 0.2, 0.8, 0.9)
concentration <- as.numeric(simulate_concave(1, 10, par))

# Adding noise and changing scale
concentration <- concentration * 18 + 2
abs.sd <- mean(concentration) * 0.1
concentration <- concentration + rnorm(10, 0, abs.sd)
concentration[concentration < 0] <- 0

# Original sampling time
time <- seq(0, 9, length.out=10) * 24

# New time for smoothing
new.time <- seq(0, 9, length.out=100) * 24

# Smoothing
corrected <- met_smooth_gam(time, concentration, new.time = new.time, k = 5)

# Plotting
plot(time, concentration, type='p',
     xlab='Time post inoculation (hours)', ylab='Concentration (mM)')
lines(new.time, corrected, type='l')
}

