% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correction.R
\name{met_smooth_loess}
\alias{met_smooth_loess}
\title{LOESS Smoothing}
\usage{
met_smooth_loess(time, concentration, new.time = NULL, ...)
}
\arguments{
\item{time}{Time or sample number for metabolite time-course.}

\item{concentration}{Metabolite concentration.}

\item{new.time}{Optional vector for new independent variables.}

\item{...}{Arguments to be passed into the loess() smoothing function,
such as span -- "the parameter alpha which controls the degree of
smoothing". If no additional arguments are provided, a reasonable
default for span (0.75) is assumed.}
}
\value{
A vector of smoothed concentrations.
}
\description{
Wraps LOESS smoothing function for use in bias correction algorithm.
WARNING: LOESS is not a particularly good choice for bias correction.
This function is included for completeness only.
}
\examples{
# Simulating concave curve with 10 points
par <- c(3.5, 4.5, 2.5, 3.5, 0.0, 0.2, 0.8, 0.9)
concentration <- as.numeric(simulate_concave(1, 10, par))

# Adding noise and changing scale
concentration <- concentration * 18 + 2
abs.sd <- mean(concentration) * 0.1
concentration <- concentration + rnorm(10, 0, abs.sd)
concentration[concentration < 0] <- 0

# Original sample time
time <- seq(0, 9, length.out=10) * 24

# New time for smoothing
new.time <- seq(0, 9, length.out=100) * 24

# Smoothing
corrected <- met_smooth_loess(time, concentration,
                              new.time = new.time, span = 0.75)

# Plotting
plot(time, concentration, type='p',
     xlab='Time post inoculation (hours)', ylab='Concentration (mM)')
lines(new.time, corrected, type='l')
}

