% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{rcon}
\alias{rcon}
\title{Constrains Random Samples}
\usage{
rcon(n, rdist, con, max.iter = 100)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{rdist}{User-defined function for generating random samples
(must take \code{n} as only argument).}

\item{con}{Vector of lower and upper constraints on generated samples.}

\item{max.iter}{Maximum number of iterations before aborting.}
}
\value{
A vector of \code{n} values sampled from \code{rdist} such that
        the minimum sampled value is greater than or equal to \code{con[1]}
        and the the maximum is smaller than or equal to \code{con[2]}.
}
\description{
Resamples supplied distribution to ensure generated values fall
within given constraints.
}
\examples{
# Using default distribution parameters
out <- rcon(10000, rnorm, c(-1, 1))
print(summary(out))
hist(out, 20, xlim=c(-2, 2), probability=TRUE,
     main='', xlab='Random variable value')

# Modifying mean and standard deviation
out <- rcon(10000, function(n) {rnorm(n, 0.5, 0.5)}, c(-1, 1))
print(summary(out))
hist(out, 20, xlim = c(-2, 2), probability = TRUE,
     main = '', xlab = 'Random variable value')
}

