% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_characteristic}
\alias{simulate_characteristic}
\title{Simulate Curve Characteristic}
\usage{
simulate_characteristic(n, dist1, par1, dist2 = NULL, par2 = NULL,
  p = 0.5, con = NULL)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{dist1}{Distribution function for the first distribution e.g. rnorm}

\item{par1}{A list of parameters for the first distribution (to be passed
into dist1 using do.call).}

\item{dist2}{Optional distribution function for the second distribution.
Note, \code{par2} has to be specified if \code{dist2} is
specified.}

\item{par2}{An optional list of parameters for the second distribution.
Note, \code{dist2} has to be specified if \code{par2} is
specified.}

\item{p}{Proportion of samples to draw using \code{dist1} vs. \code{dist2}.
Note, \code{p} is ignored if \code{dist2} or \code{par2}
is not specified.}

\item{con}{Optional vector of lower and upper constraints on generated
samples.}
}
\value{
A vector of randomly generated values.
}
\description{
Underlying function for simulating maximum concentration, relative
changes in concentration and relative standard deviations. Each of these
simulation relies on the mixture of two distributions, with the nature
of the distributions dependent on the characteristic to be simulated.
}

