% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_linear}
\alias{simulate_linear}
\title{Simulate linear trends}
\usage{
simulate_linear(n, n.samples, p)
}
\arguments{
\item{n}{Number of trends to generate.}

\item{n.samples}{Number of timepoints within each trend.}

\item{p}{Proportion of trends decreasing (vs. increasing).}
}
\value{
A dataframe with each row representing a metabolite concentration
        time-course scaled between 0 and 1. The corresponding x variables 
        are assumed to be equally spaced between 0 and 1 i.e. 
        x <- seq(0, 1, length.out=n.samples).
}
\description{
Simulates linear trends with p fraction decreasing and (1-p) fraction
increasing.
}
\examples{
trends <- simulate_linear(1000, 100, 0.75)

# Conversion for plotting
y_mat <- t(as.matrix(trends))
x <- seq(0, 1, length.out=100)
matplot(x, y_mat, type = 'l', lty = 1, lwd = 4, col = grey(0, 0.05),
        xlab = 'Relative culturing time', ylab = 'Relative concentration')
}
