% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_trend}
\alias{simulate_trend}
\title{Simulate trends}
\usage{
simulate_trend(n, n.samples, f_trend, par1, par2 = NULL, p = 0.5)
}
\arguments{
\item{n}{Number of trends to generate.}

\item{n.samples}{Number of timepoints within each trend.}

\item{f_trend}{Function that taxes x values and a vector of parameters
and generates a corresponding set of y values.}

\item{par1}{Parameter vector of lower and upper bounds for first set of 
trends.}

\item{par2}{Optional parameter vector of lower and upper bounds for second 
set of trends.}

\item{p}{Proportion of trends to generate using \code{par1} vs. \code{par2}.
Note, \code{p} is ignored if \code{par2} is not specified.}
}
\value{
A dataframe with each row representing a metabolite concentration
        time-course scaled between 0 and 1. The corresponding x variables 
        are assumed to be equally spaced between 0 and 1 i.e. 
        x <- seq(0, 1, length.out=n.samples).
}
\description{
Underlying function for simulating different types of time-course trends.
Each trend supports parameter generation from one or two sets of parameter
limits.
}
