% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{stretch}
\alias{stretch}
\title{Scales random samples}
\usage{
stretch(x, con)
}
\arguments{
\item{x}{Vector of values to stretch.}

\item{con}{Vector in the form of c(min, max) specifying new constraints}
}
\value{
A vector of values rescaled to have a minimum of \code{con[1]} and
        a maximum of \code{con[2]}.
}
\description{
Stretches a vector of values between specified constraints.
}
\examples{
out <- stretch(rnorm(10000), c(-2, 1))
print(summary(out))
hist(out, 20, xlim=c(-2, 2), probability=TRUE, 
     main = '', xlab = 'Random variable value')
}
