% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metcourse.R
\docType{data}
\name{timecourse}
\alias{timecourse}
\title{A set of realistic metabolic time-courses.}
\format{A data frame with 400 rows and 5 variables. The data simulates 1
        metabolomic experiment where 40 metabolites are tracked over 10
        timepoints.
\describe{
   \item{experiment}{Experiment number (a column of 1s in this case)}
   \item{metabolite}{Metabolites numbered 1 to 40}
   \item{sample}{Samples numbered 1 to 10}
   \item{concentration}{Metabolite concentrations (in mM)}
   \item{time}{Time post inoculation}
}}
\usage{
timecourse
}
\description{
A set of metabolic time-courses designed to mimic a suspension culture of
insect cells. The parameters used to generate this data can be seen in
`timecourse_param`.
}
\keyword{datasets}

