% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metcourse.R
\docType{data}
\name{timecourse_param}
\alias{timecourse_param}
\title{Parameters used to simulate a set of realistic metabolic time-courses.}
\format{A list with 24 parameters. See example of `simulate_timecourse`
        for more information.}
\usage{
timecourse_param
}
\description{
Parameters designed to reflect commonly observed time-courses from a
suspension culture of insect cells.
}
\keyword{datasets}

