% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{correct_rel_bias}
\alias{correct_rel_bias}
\title{Correct systematic bias in metabolomic time-course data}
\usage{
correct_rel_bias(
  time,
  compound,
  concentration,
  min.deviation = NULL,
  degree = 3,
  knots = c(0.5)
)
}
\arguments{
\item{time}{A vector of times or sample numbers for metabolite time-courses.
Note, there should be a time value for every \code{concentration} 
for every \code{compound} e.g. time = c(1, 2, 3, 1, 2, 3), 
concentration = c(20, 15, 10, 3, 6, 9), 
compound = c('glc', 'glc', 'glc', 'lac', 'lac', 'lac').}

\item{compound}{A vector of metabolite names that correspond to \code{time}
and \code{concentration}.}

\item{concentration}{A vector of metabolite concentrations.}

\item{min.deviation}{Smallest median relative deviation to identify as a 
bias.}

\item{degree}{degree of the B-spline}

\item{knots}{position of the knots of the B-spline as a fraction. 
e.g. knots <- c(0.25, 0.5, 0.75)}
}
\value{
A dataframe with the fit.
}
\description{
Identifies systematic deviations in metabolic data using a spline fit.
Timepoints that have a median relative deviation above a min.deviation value
are assumed to be influenced by a measurement or methodological bias as
compared to the overall trends metabolite concentrations.
}
\examples{

# Using previously simulated data 40 metabolic trends with 10 time points
data(timecourse)

# Adding an error of 5\% at sample 4
logic <- timecourse$sample == 4
timecourse$concentration[logic] <- timecourse$concentration[logic] * 1.05 

# Correcting
output <- correct_rel_bias(timecourse$time, 
                                         timecourse$concentration,
                                         timecourse$metabolite)
timecourse$corrected <- output$fit

# Plotting -- the original value of the corrected point is marked in red
par(mfrow = c(8, 5), oma = c(5, 4, 1, 1) + 0.1, mar = c(1, 1, 1, 1) + 0.1)
new.time <- seq(min(timecourse$time), max(timecourse$time), length.out=100)

for (metabolite in unique(timecourse$metabolite)) {

  logic <- timecourse$metabolite == metabolite
  d <- timecourse[logic, ]
  
  logic2 <- d$concentration != d$corrected

  plot(d$time, d$corrected, pch = 16, xlab = '', ylab = '',  
       ylim = c(min(d$concentration), max(d$concentration)))

  smoothed <- met_smooth_gam(d$time, d$corrected,
                             new.time = new.time, k = 5)
  lines(new.time, smoothed)

  points(d$time[logic2], d$concentration[logic2], pch = 16, col = 'red')

}

title(xlab = 'Time post inoculation (hours)', 
      ylab = 'Concentration (mM)', outer = TRUE, line = 3)
}
