% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_concave}
\alias{simulate_concave}
\title{Simulating concave trends}
\usage{
simulate_concave(n, n.samples, p, par1, par2)
}
\arguments{
\item{n}{Number of trends to generate.}

\item{n.samples}{Number of timepoints within each trend.}

\item{p}{Proportion of trends to generate using \code{par1} vs. \code{par2}.}

\item{par1}{Parameter vector (a_min, a_max, b_min, b_max, c_min, c_max,
d_min, d_max) for one set of trends where y = x**(a-1)*(1-x)**(b-1)
constrained to min(x) = c, max(x) = d}

\item{par2}{Parameter vector (a_min, a_max, b_min, b_max, c_min, c_max,
d_min, d_max) for one set of trends where y = x**(a-1)*(1-x)**(b-1)
constrained to min(x) = c, max(x) = d}
}
\value{
A dataframe with each row representing a metabolite concentration
        time-course scaled between 0 and 1. The corresponding x variables
        are assumed to be equally spaced between 0 and 1 i.e.
        x <- seq(0, 1, length.out=n.samples).
}
\description{
Simulates concave trends by mixing trimmed and scaled beta distributions.
parameters are sampled from uniform distributions with boundaries provided
as input to this function.
}
\examples{
par1 <- c(3.5, 4.5, 2.5, 3.5, 0.0, 0.2, 0.8, 0.9)
par2 <- c(3.5, 4.5, 2.5, 3.5, 0.2, 0.4, 0.7, 0.8)
trends <- simulate_concave(1000, 100, 0.75, par1, par2)

# Conversion for plotting
y_mat <- t(as.matrix(trends))
x <- seq(0, 1, length.out=100)
matplot(x, y_mat, type = 'l', lty = 1, lwd = 4, col = grey(0, 0.05),
        xlab = 'Relative culturing time', ylab = 'Relative concentration')
}

