% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_decreasing}
\alias{simulate_decreasing}
\title{Simulating decreasing trends}
\usage{
simulate_decreasing(n, n.samples, p, par1, par2)
}
\arguments{
\item{n}{Number of trends to generate.}

\item{n.samples}{Number of timepoints within each trend.}

\item{p}{Proportion of trends to generate using \code{par1} vs. \code{par2}.}

\item{par1}{Parameter vector (a_min, a_max, b_min, b_max) for one set of
trends where y = (1 + exp((x-a)/b))**(-1)}

\item{par2}{Parameter vector (a_min, a_max, b_min, b_max) for second set of
trends where y = (1 + exp((x-a)/b))**(-1)}
}
\value{
A dataframe with each row representing a metabolite concentration
        time-course scaled between 0 and 1. The corresponding x variables
        are assumed to be equally spaced between 0 and 1 i.e.
        x <- seq(0, 1, length.out=n.samples).
}
\description{
Simulates decreasing trends by mixing two sigmoid curve families. Sigmoid
parameters are sampled from uniform distributions with boundaries provided
as input to this function.
}
\examples{
par1 <- c(0.2, 0.6, 0.10, 0.18)
par2 <- c(0.6, 0.9, 0.10, 0.18)
trends <- simulate_decreasing(1000, 100, 0.05, par1, par2)

# Conversion for plotting
y_mat <- t(as.matrix(trends))
x <- seq(0, 1, length.out=100)
matplot(x, y_mat, type = 'l', lty = 1, lwd = 4, col = grey(0, 0.05),
        xlab = 'Relative culturing time', ylab = 'Relative concentration')
}

