% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_max}
\alias{simulate_max}
\title{Simulates Maximum Concentrations}
\usage{
simulate_max(n, par1, par2 = NULL, p = 0.5, con = NULL)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{par1}{Parameters (mean, sd) in concentration units for first
distribution. These values are converted to logarithms for
sampling.}

\item{par2}{Optional parameters (mean, sd) in concentration units for second
distribution. These values are converted to logarithms for
sampling.}

\item{p}{Proportion of samples to draw using \code{par1} vs. \code{par2}.
Note, \code{p} is ignored if \code{par2} is not specified.}

\item{con}{Optional vector of lower and upper constraints on generated
samples.}
}
\value{
A vector of concentration values.
}
\description{
Simulates maximum metabolite concentrations using a mixture of 2 normal
distributions of concentration logarithms.
}
\examples{
# Generating concentrations
out <- simulate_max(10000, c(7, 2), c(0.5, 2), 0.3)

# Formatting output on logarithmic scale
out <- log10(out)
labels <- c(0.1, 0.25, 0.5, 1, 2.5, 5, 10, 25, 50)
hist(out, 20, axes = FALSE, probability = TRUE,
     main = '', xlab = 'Maximum metabolite concentration')
axis(side = 2)
axis(at = log10(labels), labels = labels, side = 1)
}

