% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_rel_change}
\alias{simulate_rel_change}
\title{Simulating concentration changes}
\usage{
simulate_rel_change(n, p, par1, par2, con = NULL)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{p}{Proportion of samples to draw using \code{par1} vs. \code{par2}.}

\item{par1}{Parameters (alpha, beta) for first distribution.}

\item{par2}{Parameters (alpha, beta) for second distribution.}

\item{con}{Optional vector of lower and upper constraints on generated
samples.}
}
\value{
A vector of relative concentration values.
}
\description{
Simulates relative (fractional) changes in metabolite concentrations using
a mixture of beta distributions.
}
\examples{
out <- simulate_rel_change(10000, 0.7, c(2, 5), c(0.5, 0.5), c(0.1, 1))
hist(out, 20)
}

